+++ aiiabox/apps/chats/views/__init__.py
@@ -1,0 +1,6 @@
+from .create import ChatCreateView
+from .delete import ChatDeleteView
+from .detail import ChatDetailView
+from .list import ChatListView
+
+__all__ = ["ChatListView", "ChatDetailView", "ChatCreateView", "ChatDeleteView"]